test: BookService 서비스 생성에 따른 테스트 코드 분리
