+++ src/test/java/roomescape/fixture/FakeReservationRepositoryFixture.java
@@ -1,0 +1,20 @@
+package roomescape.fixture;
+
+import java.time.LocalDate;
+import java.time.LocalTime;
+import java.util.List;
+import roomescape.domain.Reservation;
+import roomescape.domain.ReservationTime;
+import roomescape.domain.Theme;
+import roomescape.repository.FakeReservationRepository;
+
+public class FakeReservationRepositoryFixture {
+
+    public static FakeReservationRepository create() {
+        return new FakeReservationRepository(List.of(
+                new Reservation(1L, "가이온", LocalDate.now().plusDays(7),
+                        new ReservationTime(1L, LocalTime.of(10, 0)),
+                        new Theme(1L, "우테코", "방탈출", "https://"))
+        ));
+    }
+}
